split core tests into separate files
